requirements.txt
pyaudio
websocket-client
python-dotenv
ibm-cloud-sdk-core
json